Allow Frontend or API to be disabled

+++ app/Http/routes.php
@@ -6,12 +6,14 @@
 |--------------------------------------------------------------------------
 */
 
-$app->get('/', ['as' => 'index', 'uses' => 'FrontendController@getIndex']);
+if(env('APP_FRONTEND')) {
+	$app->get('/', ['as' => 'index', 'uses' => 'FrontendController@getIndex']);
 
-$app->post('/generate', ['as' => 'generate', 'middleware' => 'csrf', 'uses' => 'BackendController@postGenerate']);
-$app->post('/destroy', ['as' => 'destroy', 'middleware' => 'csrf', 'uses' => 'BackendController@postDestroy']);
+	$app->post('/generate', ['as' => 'generate', 'middleware' => 'csrf', 'uses' => 'BackendController@postGenerate']);
+	$app->post('/destroy', ['as' => 'destroy', 'middleware' => 'csrf', 'uses' => 'BackendController@postDestroy']);
+}
 
-$app->post('/api/generate', ['as' => 'generate', 'uses' => 'BackendController@apiPostGenerate']);
-$app->delete('/api/destroy', ['as' => 'destroy', 'uses' => 'BackendController@apiDeleteDestroy']);
-
-
+if(env('APP_API')) {
+	$app->post('/api/generate', ['as' => 'generate', 'uses' => 'BackendController@apiPostGenerate']);
+	$app->delete('/api/destroy', ['as' => 'destroy', 'uses' => 'BackendController@apiDeleteDestroy']);
+}
